fix: Reduce and improve the PR template (#3)
Since we are using conventional commit messages including for PR titles,
the "type of change" section is redundant.

In many pull requests the template was never updated/used and irrelevant
sections were not dropped.

- By clearly describing what to write it hopefully has the same effect
of helping people to tink about it.
- By using the words **REPLACE** and **DROP** in each section, the hope
is that people will follow the instructions more often. Since the text
doesn't make any sense as part of a real PR description.
- It would be possible to add an automated check that requires people to
at least drop this text before a PR can land.

By adding a link to the file this template comes from, more people can
be aware without anybody telling them, so they are encouraged to
contribute.