feat: add git-hooks/no-commit-to-default-branch
After some research it seems to be the status quo to configure all hooks inside the repository,
instead of on the user level.
This has the advantage of knowing exactly what the default branch is,
It is also easier for everybody to be aligned with the hooks and contribute to them.

There are two main options: `husky` for npm-based projects and `pre-commit` for everything else.
So I documented both options.